+++ ios/HYDrawing/HYDrawing/HYBrushCore.mm
@@ -36,15 +36,17 @@
 /// 初始化
 - (BOOL) initializeWithWidth:(float)w height:(float)h scale:(float)s
 {
+    /// CZActiveState Initializaiton comes first, for it will help other initial work
+    CZActiveState::getInstance()->setEraseMode(false);
+    CZActiveState::getInstance()->setActiveBrush(kPencil);
+    CZActiveState::getInstance()->mainScreenScale = s;
+    
     viewImpl = new CZViewImpl(CZRect(0,0,w,h));
     canvas = new CZCanvas(viewImpl);
     
     [[PaintingManager sharedInstance] initializeWithWidth:w height:h scale:s];
     painting = (CZPainting*)[[PaintingManager sharedInstance] getInitialPainting];
     canvas->setPaiting(painting);
-    CZActiveState::getInstance()->setEraseMode(false);
-    CZActiveState::getInstance()->setActiveBrush(kPencil);
-    CZActiveState::getInstance()->mainScreenScale = s;
     [self setActiveBrushwatercolorPenStamp ];
     
     self.hasInitialized = YES;